Add initial rough building process
This lays out groundwork for the building process. This process is
based on streams, similar to gulpjs. I didn't find much to build on
from existing python libraries, so I created a basic one myself. It has
a concept of sources, pipes and sinks, through which liquids (files)
run. The idea is the same as piping on the command-line, but only
liquids can be passed around (but a pipe may return entirely different
liquids from what it received).

More on this probably if I write docs at some point.